Hopefully dynamic pick working well
I undid the separation of dynamic grab into two functions.
The issue with the way is was being done is that if no items are in the
queue it will wait forever. There needed to be a way to cancel when
the user presses stop.

To allow for canceling the dynamic grab I added a `runningStopped` event

This function will wait for the next item but will cancel if
`runningStopped` is emitted.

```
const getNextItem = () => Observable
  .interval(50)
  .takeUntil(runningStopped)
  .map(() => queue.getClosestItemToRobot())
  .filter(item => item !== undefined)
  .take(1)
  .toPromise();
```

Additionally having the robot wait for an item needs to be cancelable.
If the belt is stopped the item will never reach it and it will be
blocked forever.

To allow for this the `runningStopped` event is passed to the
`dynamicGrab` function. Here it will keep taking the item coordinates
until it is in range or `runningStopped` is emitted or the item was
destroyed by someone else (queue cleared).

In the case where the item moves in range the code completes normally.
In the other two cases however `first()` will emit an error because
it never was invoked. The try-catch will cause the dynamicGrab to exit
early.

```
try {
  await item.coordsUpdated
    .takeUntil(runningStopped)
    .map(() => predictTarget())
    .do(targ => console.log(targ))
    .filter(targ => targ.x >= this.cal.minPick.x)
    .first()
    .toPromise();
} catch {
  item.destroy();
  return;
}
```